Issue #271: Factorise SNMP and SNMPV3 Extensions * Factorise check protocol * Add Java doc
